+++ model_navigator/commands/infer_metadata.py
@@ -80,13 +80,14 @@ def _extract_max_batch_size(axes_shapes: Dict[str, Dict[int, List[int]]], batch_
     return 0
 
 
-def _get_trt_profile_from_axes_shapes(axes_shapes, batch_dim):
+def _get_trt_profile_from_axes_shapes(axes_shapes, batch_dim, max_batch_size=None):
     trt_profile = TensorRTProfile()
     for name, axes in axes_shapes.items():
         min_max_opt = []
         for ax, shapes in axes.items():
             if ax == batch_dim:  # min bs = 1
-                min_max_opt.append((1, int(np.median(shapes)), max(shapes)))
+                max_batch_size = max_batch_size or max(shapes)
+                min_max_opt.append((1, int(np.median(shapes)), max_batch_size))
             else:
                 min_max_opt.append((min(shapes), int(np.median(shapes)), max(shapes)))
         if min_max_opt: